weatherStateをAlarmUiStateに追加
