+++ app/src/main/java/com/example/weatheralarmapp/ui/features/alarm/components/ToggleTimePicker.kt
@@ -38,6 +38,7 @@ import androidx.compose.ui.window.DialogProperties
 import com.example.weatheralarmapp.R
 import com.example.weatheralarmapp.ui.features.alarm.AlarmItemState
 import com.example.weatheralarmapp.ui.features.alarm.AlarmUiState
+import com.example.weatheralarmapp.ui.features.alarm.WeatherState
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -159,6 +160,7 @@ private fun DialTimePickerPreview() {
                         isWeatherForecastOn = false,
                     ),
                 expandedAlarmItem = false,
+                weatherState = WeatherState.Initial,
             ),
     )
 }
